@@ -1,15 +1,26 @@
-import React, {ChangeEvent, KeyboardEvent, useState} from "react";
-import {IconButton, TextField} from "@material-ui/core";
-import {AddCircleOutline} from "@material-ui/icons";
+import React, { ChangeEvent, KeyboardEvent, useState } from 'react';
+import TextField from '@mui/material/TextField';
+import IconButton from '@mui/material/IconButton';
+import { AddBox } from '@mui/icons-material';
 
-type AddItemFormType = {
+type AddItemFormPropsType = {
     addItem: (title: string) => void
 }
 
-export const AddItemForm = React.memo((props: AddItemFormType) => {
-    console.log('AddItemForms is called')
-    let [title, setTitle] = useState("")
+export const AddItemForm = React.memo(function (props: AddItemFormPropsType) {
+    console.log('AddItemForm called')
+
+    let [title, setTitle] = useState('')
     let [error, setError] = useState<string | null>(null)
+
+    const addItem = () => {
+        if (title.trim() !== '') {
+            props.addItem(title);
+            setTitle('');
+        } else {
+            setError('Title is required');
+        }
+    }
 
     const onChangeHandler = (e: ChangeEvent<HTMLInputElement>) => {
         setTitle(e.currentTarget.value)
@@ -20,33 +31,21 @@ export const AddItemForm = React.memo((props: AddItemFormType) => {
             setError(null);
         }
         if (e.charCode === 13) {
-            addTask();
+            addItem();
         }
     }
 
-    const addTask = () => {
-        if (title.trim() !== "") {
-            props.addItem(title.trim());
-            setTitle("");
-        } else {
-            setError("Title is required");
-        }
-    }
-
-
-    return <div style={{textAlign: 'center'}}>
-        <TextField value={title}
-                   label={'add value'}
+    return <div>
+        <TextField variant="outlined"
+                   error={!!error}
+                   value={title}
                    onChange={onChangeHandler}
                    onKeyPress={onKeyPressHandler}
-                   error={!!error}
+                   label="Title"
                    helperText={error}
         />
-        <IconButton onClick={addTask}
-                    color={"primary"}
-                    size={"small"}>
-            <AddCircleOutline/>
+        <IconButton color="primary" onClick={addItem}>
+            <AddBox/>
         </IconButton>
     </div>
-
-});
+})
