@@ -1,18 +1,21 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import * as serviceWorker from './serviceWorker';
-import {AppWithRedux} from './AppWithRedux'
-import {Provider} from "react-redux";
-import {store} from "./store";
+import reportWebVitals from './reportWebVitals';
+import App from './App';
+import {Provider} from 'react-redux';
+import {store} from './state/store';
 
 ReactDOM.render(
-    <Provider store={store}>
-        <AppWithRedux/>
-    </Provider>
-    , document.getElementById('root'));
+    <React.StrictMode>
+        <Provider store={store}>
+            <App/>
+        </Provider>
+    </React.StrictMode>,
+    document.getElementById('root')
+);
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
