@@ -1,99 +1,106 @@
-import {TasksStateType} from "../App";
-import {v1} from "uuid";
-import {AddTodoListActionType, RemoveTodoListActionType} from "./todolists-reducer";
-
-type  ActionsType = RemoveTaskActionType | AddTaskActionType
-    | ChangeTaskStatusType | changeTaskTitleType
-    | AddTodoListActionType | RemoveTodoListActionType
+import { TaskType } from '../Todolist';
+import { v1 } from 'uuid';
+import { AddTodolistActionType, RemoveTodolistActionType } from './todolists-reducer';
+import { TasksStateType } from '../App';
 
 export type RemoveTaskActionType = {
-    type: "REMOVE-TASK"
-    todoListID: string
-    id: string
+    type: 'REMOVE-TASK',
+    todolistId: string
+    taskId: string
+}
 
-};
 export type AddTaskActionType = {
-    type: 'ADD-TASK'
-    title: string
-    todoListID: string
-}
-export type ChangeTaskStatusType = {
-    type: 'CHANGE-TASK-STATUS'
-    id: string
-    taskStatus: boolean
-    todoListID: string
-}
-export type changeTaskTitleType = {
-    type: 'CHANGE-TASK-TITLE'
-    todoListID: string
-    id: string
+    type: 'ADD-TASK',
+    todolistId: string
     title: string
 }
 
-export const initialState : TasksStateType= { }
-// {[todoListID1]: [
-//         {id: v1(), title: "HTML&CSS", isDone: true},
-//         {id: v1(), title: "JS", isDone: true},
-//         {id: v1(), title: "ReactJS", isDone: false},
-//         {id: v1(), title: "Rest API", isDone: false},
-//         {id: v1(), title: "GraphQL", isDone: false},
-//     ], [todoListID2]: [
-//         {id: v1(), title: "HTML&CSS", isDone: true},
-//         {id: v1(), title: "JS", isDone: true},
-//         {id: v1(), title: "ReactJS", isDone: false},
-//         {id: v1(), title: "Rest API", isDone: false},
-//         {id: v1(), title: "GraphQL", isDone: false},
-//     ]
-// }
+export type ChangeTaskStatusActionType = {
+    type: 'CHANGE-TASK-STATUS',
+    todolistId: string
+    taskId: string
+    isDone: boolean
+}
 
-export const tasksReducer = (state: TasksStateType = initialState, action: ActionsType) => {
+export type ChangeTaskTitleActionType = {
+    type: 'CHANGE-TASK-TITLE',
+    todolistId: string
+    taskId: string
+    title: string
+}
+
+type ActionsType = RemoveTaskActionType | AddTaskActionType
+    | ChangeTaskStatusActionType
+    | ChangeTaskTitleActionType
+    | AddTodolistActionType
+    | RemoveTodolistActionType
+
+const initialState: TasksStateType = {}
+
+export const tasksReducer = (state: TasksStateType = initialState, action: ActionsType): TasksStateType => {
     switch (action.type) {
         case 'REMOVE-TASK': {
-            let copyState = {...state}
-            let todoListTask = copyState[action.todoListID]
-            copyState[action.todoListID] = todoListTask.filter(tl => tl.id !== action.id)
-            return copyState
+            const stateCopy = {...state}
+            const tasks = stateCopy[action.todolistId];
+            const newTasks = tasks.filter(t => t.id !== action.taskId);
+            stateCopy[action.todolistId] = newTasks;
+            return stateCopy;
         }
         case 'ADD-TASK': {
-            let copyState = {...state}
-            let task = {id: v1(), title: action.title, isDone: false}
-            let todoListTasks = copyState[action.todoListID]
-            copyState[action.todoListID] = [task, ...todoListTasks]
-            return copyState
+            const stateCopy = {...state}
+            const newTask: TaskType = {
+                id: v1(),
+                title: action.title,
+                isDone: false
+            }
+            const tasks = stateCopy[action.todolistId];
+            const newTasks = [newTask, ...tasks];
+            stateCopy[action.todolistId] = newTasks;
+            return stateCopy;
         }
         case 'CHANGE-TASK-STATUS': {
-            let task = state[action.todoListID]
-            state[action.todoListID] = task.map(t => t.id === action.id ? {...t, isDone: action.taskStatus}: t);
-            return {...state}
-        }
-        case "CHANGE-TASK-TITLE": {
-            let task = state[action.todoListID]
-            state[action.todoListID] = task.map(t => t.id === action.id ? {...t,  title: action.title } : t)
-            return {...state}
-        }
-        case "ADD-TODOLIST":{
-            return {...state, [action.todoListId]: []}
-        }
-        case "REMOVE-TODOLIST":{
-            let stateCopy = {...state}
-            delete stateCopy[action.id]
-            return stateCopy
-        }
+            let todolistTasks = state[action.todolistId];
+            let newTasksArray = todolistTasks
+                .map(t => t.id === action.taskId ? {...t, isDone: action.isDone} : t);
 
+            state[action.todolistId] = newTasksArray;
+            return ({...state});
+        }
+        case 'CHANGE-TASK-TITLE': {
+            let todolistTasks = state[action.todolistId];
+            // найдём нужную таску:
+            let newTasksArray = todolistTasks
+                .map(t => t.id === action.taskId ? {...t, title: action.title} : t);
+
+            state[action.todolistId] = newTasksArray;
+            return ({...state});
+        }
+        case 'ADD-TODOLIST': {
+            return {
+                ...state,
+                [action.todolistId]: []
+            }
+        }
+        case 'REMOVE-TODOLIST': {
+            const copyState = {...state};
+            delete copyState[action.id];
+            return copyState;
+        }
         default:
             return state;
     }
 }
 
-export const removeTaskAC = (id: string, todoListID: string): RemoveTaskActionType => {
-    return {type: 'REMOVE-TASK', todoListID, id}
+export const removeTaskAC = (taskId: string, todolistId: string): RemoveTaskActionType => {
+    return {type: 'REMOVE-TASK', taskId: taskId, todolistId: todolistId}
 }
-export const addTaskAC = (title: string, todoListID: string): AddTaskActionType => {
-    return {type: "ADD-TASK", title, todoListID}
+export const addTaskAC = (title: string, todolistId: string): AddTaskActionType => {
+    return {type: 'ADD-TASK', title, todolistId}
 }
-export const changeTaskStatusAC = (id: string, taskStatus: boolean, todoListID: string): ChangeTaskStatusType => {
-    return {type: "CHANGE-TASK-STATUS", id, taskStatus, todoListID}
+export const changeTaskStatusAC = (taskId: string, isDone: boolean, todolistId: string): ChangeTaskStatusActionType => {
+    return {type: 'CHANGE-TASK-STATUS', isDone, todolistId, taskId}
 }
-export const changeTaskTitleAC = (todoListID: string, id: string, title: string): changeTaskTitleType => {
-    return {type: 'CHANGE-TASK-TITLE', todoListID, id, title}
+export const changeTaskTitleAC = (taskId: string, title: string, todolistId: string): ChangeTaskTitleActionType => {
+    return {type: 'CHANGE-TASK-TITLE', title, todolistId, taskId}
 }
+
